create_dataset/requirements.txt
openai
pydantic
tqdm
python-slugify
python-dotenv
